Throw an Exception on Abnormal Exits in Child Processes
Previously we treated everything abnormal as a simple failure. Which can
hide a lot of issues -- like a bunch of segfaults that are super hard to
diagnose! Instead we now throw an exception. Effectively this has the
same result: the job is failed (and maybe retried), but we get an actual
error message about the exit/stop code or a wait status value if things
fail.