Introduce Driver::release
This is a way for a driver's client to release a message back to a ready
state without going through a retry or anything like that.

The reasoning here is that there are cases when releasing a message is
the correct thing to do (like if a consummer gets a SIGTERM or something
like that).