Introduce `MessageLifecycle`
This will provide a way for end users to hook into the flow of a message
through the queue without tying the core library to a specific way that
this happens. Previously (1.X) we had the symfony event system, for example,
but it was not great and we never used it.